requirements.txt
requests
python-dotenv
pydantic
PyYAML
duckdb